Renaming simple test working
